v6/requirements.txt
pytorch-lightning==1.9.5
deepspeed
bitsandbytes
wandb
prompt_toolkit
pandas
rwkv
